+++ css/tailwind.css
@@ -56,8 +56,4 @@ html {
   }
 }
 
-.glance iframe {
-  max-width: 100%;
-}
-
 @tailwind utilities;
